+++ ffi_common/src/error.rs
@@ -70,7 +70,7 @@ pub extern "C" fn get_last_err_msg() -> *const c_char {
         msg = last_error.borrow().clone();
     });
     match msg {
-        Some(string) => try_or_set_error!(CString::new(string).map(|c| c.into_raw())),
+        Some(string) => try_or_set_error!(CString::new(string).map(CString::into_raw)),
         None => std::ptr::null(),
     }
 }