+++ ffi_common/src/lib.rs
@@ -23,6 +23,7 @@
 )]
 #![forbid(missing_docs, unused_extern_crates, unused_imports)]
 
+pub mod codegen_helpers;
 #[macro_use]
 pub mod error;
 pub mod datetime;
@@ -31,4 +32,4 @@ pub mod string;
 
 use paste::paste;
 
-declare_value_type_array_struct!(bool, u8, u16, u32, u64, i8, i16, i32, i64, f32, f64);
+declare_value_type_ffi!(bool, u8, u16, u32, u64, i8, i16, i32, i64, f32, f64);