+++ ffi_derive/build.rs
@@ -1,0 +1,3 @@
+// This has to exist for Cargo to set the `OUT_DIR` environment variable at compile-time.
+
+fn main() {}